+++ src/main/java/net/ccbluex/liquidbounce/ui/client/clickgui/style/styles/panel/BoolElement.kt
@@ -6,7 +6,7 @@
 package net.ccbluex.liquidbounce.ui.client.clickgui.style.styles.panel
 
 import net.ccbluex.liquidbounce.config.BoolValue
-import net.ccbluex.liquidbounce.ui.font.Fonts.fontRegular30
+import net.ccbluex.liquidbounce.ui.font.Fonts.fontRegular35
 import net.ccbluex.liquidbounce.ui.client.clickgui.style.styles.PanelStyle.highlightColor
 import net.ccbluex.liquidbounce.ui.client.clickgui.style.styles.PanelStyle.highlightColorAlpha
 import net.ccbluex.liquidbounce.ui.client.clickgui.style.styles.PanelStyle.referenceColor
@@ -22,8 +22,8 @@ class BoolElement(
 ) : ValueElement() {
 
     override var margin: Float = 5f
-    override var height: Float = fontRegular30.fontHeight.toFloat() + margin
-    override var width: Float = fontRegular30.getStringWidth(valueName).toFloat()
+    override var height: Float = fontRegular35.fontHeight.toFloat() + margin
+    override var width: Float = fontRegular35.getStringWidth(valueName).toFloat()
 
     private var hitboxX = 0f..0f
     private var hitboxY = 0f..0f
@@ -39,25 +39,25 @@ class BoolElement(
     override fun drawElement() {
         updateElement()
 
-        fontRegular30.drawString(
+        fontRegular35.drawString(
             valueName,
             startX,
             startY,
             WHITE.rgb
         )
 
-        var circleY = startY + fontRegular30.fontHeight / 2f - 1.5f
+        var circleY = startY + fontRegular35.fontHeight / 2f - 1.5f
         var circleX = startX + width + 10f
 
         if (value.isActive()) {
             drawCircle(
                 circleX,
                 circleY,
-                1.75f,
+                2f,
                 highlightColor
             )
         } else {
-            drawCircle(circleX, circleY, 1.75f, referenceColor)
+            drawCircle(circleX, circleY, 2f, referenceColor)
         }
     }
 
@@ -71,7 +71,7 @@ class BoolElement(
         if (previousValue != null) {
             this.startY = previousValue!!.startY + previousValue!!.height
         }
-        this.hitboxX = startX .. (startX + width+14f)
-        this.hitboxY = startY .. (startY + height-margin)
+        this.hitboxX = startX .. (startX + width + 14f)
+        this.hitboxY = startY .. (startY + height - margin)
     }
 }