+++ src/main/java/net/ccbluex/liquidbounce/ui/client/clickgui/style/styles/panel/IntElement.kt
@@ -6,7 +6,7 @@
 package net.ccbluex.liquidbounce.ui.client.clickgui.style.styles.panel
 
 import net.ccbluex.liquidbounce.config.IntValue
-import net.ccbluex.liquidbounce.ui.font.Fonts.fontRegular30
+import net.ccbluex.liquidbounce.ui.font.Fonts.fontRegular35
 import net.ccbluex.liquidbounce.ui.client.clickgui.style.styles.PanelStyle.highlightColor
 import net.ccbluex.liquidbounce.ui.client.clickgui.style.styles.PanelStyle.highlightColorAlpha
 import net.ccbluex.liquidbounce.ui.client.clickgui.style.styles.PanelStyle.referenceColor
@@ -25,8 +25,8 @@ class IntElement(
 
     override var margin: Float = 5f
 
-    override var height: Float = fontRegular30.fontHeight.toFloat() + margin
-    override var width: Float = fontRegular30.getStringWidth(valueName).toFloat()
+    override var height: Float = fontRegular35.fontHeight.toFloat() + margin
+    override var width: Float = fontRegular35.getStringWidth(valueName).toFloat()
 
     private var hitboxX = 0f..0f
     private var hitboxY = 0f..0f
@@ -35,6 +35,7 @@ class IntElement(
         if (previousValue != null) {
             startY = previousValue!!.startY + previousValue!!.height
         }
+
         this.hitboxX = startX..(startX + width + 14f)
         this.hitboxY = startY..(startY + height - margin)
     }
@@ -42,7 +43,7 @@ class IntElement(
     override fun drawElement() {
         updateElement()
 
-        fontRegular30.drawString(
+        fontRegular35.drawString(
             valueName,
             startX,
             startY,
@@ -56,7 +57,7 @@ class IntElement(
         val offsetX = 100f * progress
 
         val circleX = startX + width + 10f + offsetX
-        val circleY = startY + fontRegular30.fontHeight / 2f - 1.5f
+        val circleY = startY + fontRegular35.fontHeight / 2f - 1.5f
 
         drawRect(
             startX + width + 10f,
@@ -66,13 +67,12 @@ class IntElement(
             referenceColor
         )
 
-        //drawCircle(circleX, circleY, 3f, highlightColorAlpha.rgb)
         drawCircle(circleX, circleY, 1.75f, highlightColor)
 
-        fontRegular30.drawString(
+        fontRegular35.drawString(
             value.get().toString() + " " + (value.suffix ?: ""),
             startX + width + 120f,
-            circleY - fontRegular30.fontHeight / 4f,
+            circleY - fontRegular35.fontHeight / 4f,
             WHITE.rgb
         )
     }
@@ -81,6 +81,7 @@ class IntElement(
         if (previousValue != null) {
             this.startY = previousValue!!.startY + previousValue!!.height
         }
+
         this.hitboxX = startX + width + 10f..(startX + width + 110f)
         this.hitboxY = startY..(startY + height - margin)
     }
@@ -91,6 +92,7 @@ class IntElement(
             val max = startX + width + 110f
             val progress = (mouseX - min) / (max - min)
             var newValue = value.lerpWith(progress)
+
             // Round to 2 decimal places
             value.set(newValue)
         }